LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focusgroup/tentative/forward-navigation/wraps-and-goes-into-inner-focusgroup-expected.txt

item1 item2
item3

FAIL When the focus is set on the last item of a focusgroup that supports wrapping in the axis of the arrow key pressed and the first item is in an inner focusgroup that supports it too, the focus moves to that item in the inner focusgroup. assert_equals: expected Element node <span id="item1" tabindex="-1">item1</span> but got Element node <span id="item3" tabindex="0">item3</span>
